requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
astral==3.2
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
Bottleneck==1.5.0
cachetools==6.1.0
certifi==2025.1.31
cftime==1.6.4.post1
chaospy==4.3.20
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cplex==22.1.2.0
crc32c==2.7.1
cycler==0.12.1
dask==2025.5.1
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
docplex==2.30.251
donfig==0.8.1.post1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fonttools==4.55.8
fsspec==2025.5.1
geopandas==1.1.1
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
h11==0.16.0
h5netcdf==1.6.3
h5py==3.13.0
highspy==1.11.0
hydesign==1.5.3
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jenn==1.0.8
Jinja2==3.1.6
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightgbm==4.6.0
linopy==0.5.5
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
meteostat==1.7.4
narwhals==1.27.1
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.5
NREL-PySAM==7.0.0
numcodecs==0.16.1
numexpr==2.11.0
numpoly==1.3.6
numpy==2.2.3
numpy-financial==1.0.0
openmdao==3.39.0
openpyxl==3.1.5
orjson==3.10.18
ortools==9.14.6206
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
polars==1.31.0
prompt_toolkit==3.0.50
protobuf==6.31.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pvlib==0.11.2
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyDOE3==1.0.5
Pygments==2.19.1
pyogrio==0.11.0
pyparsing==3.2.1
pyproj==3.7.1
pypsa==0.35.1
python-dateutil==2.9.0.post0
pytz==2025.1

pyzmq==26.2.1
rainflow==3.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==80.9.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
smt==2.9.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.4
streamlit==1.46.1
tenacity==9.1.2
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
xarray==2025.6.1
xgboost==2.1.4
zarr==3.0.9
zipp==3.23.0
streamlit-autorefresh
